github_content_proxy/requirements.txt
requests==2.32.3
PyJWT==2.9.0
python-dotenv==1.0.1